Consolidate account related signal handlers
